* Fix parsing double and floats (#8)

* Core/Auras: Implemented using all aura interrupt flag fields

* DB/Misc: Make base files more read-able

* Core/Spells: Removed outdated disengage script - it can be cast out of combat

* Fixes Script loading.

Some cleanup on array Clear
Misc errors

* Core/Spells: Implemented spells causing Forbearance

* Core/Spells: Moved applying exclusive aura spells to scripts

* Core/Spell: move attribute helpers to spellInfo

* Core/Spells: Glyph of Desperation/Pain Suppression has been removed

* Core/Spell: revamp of immunities system

* Fixes core not working on windows 7.

* Core/Spells: Fixed proc cooldown

* Core/Unit: consider creatures with same faction to be on the same party/raid for the purposes of spell/aura targets

* Core/SpellScript: allow optional parameters to be passed to FinishCast

* Core/Player: Automatically learn riding spells for certain race/class combinations

* Fixes ProcChance always being a 0

* Misc fixes

* Core/Pools: PoolManager cleanup

* Core/Spells: Fixed SPELL_AURA_MOD_INCREASE_ENERGY_PERCENT

* Core/Globals: Load spellspecific and aurastate after loading spell ranks

Core/Spells: Removed obsolete code
Core/Spell: do not fail triggered Charge effects in case a root is present.

* Core/Scripts: add checks to Master's Call script

* Core/GameObject: use a friendly trigger for positive spells.

DB/Creature: correct 2 npc walk/run speed
DB/Spell: Vampiric Embrace proc
DB/GameObjects: Cursed Egg

* Core/Unit: damage immune improvements

* Core/RBAC: Cleanup

* Forgot sql files

* Core/Quests: Implemented player choices

* Core/Quests: Implemented player choice rewards

DB/Creature: Cosmetic Mottled Raptor

* Core/BattlePets: Misc fixes

* Forgot to add missing command

* Happy New Year!!!!

* Really fix parsing double and floats, Also added to BNetServer just to make sure.

* Core/Misc: Added helper function Unit::SetFullPower * Also renamed Unit::getPowerType and Unit::setPowerType to follow style guidelines

* Core/Spells: Removed obsolete spell fixes

* Core/PacketIO: Fixed SMSG_TRANSFER_PENDING structure

* Core/Packets: Implemented SMSG_GAME_OBJECT_UI_ACTION

* Fixed string getting cut off if the string had some symbols. The length of the string would not be the same as the bytes length.

* Some command fixes

* Misc cleanups

* Core/Creature: Added TitleAlt to creature_template

* Misc DB updates

* Core/Maps: Fixed objects being left in grids.

* Core/Entities: Extended SpawnMask to 64 bit

* Misc DB fixes

* Core/Items: Try to automatically equip new bags

* Core/Maps: Treat scenario maps as instanceable

* Core/Entities: Improved UpdateAreaAndZonePhase() to allow easier building of database phase conditions

* Core/Entities: Remove phasemask

* Core/SmartScripts: Fixes repeating events to soon when using waypoints

* Core/Misc: Misc fixes/Cleanups

* Core/LFG: Fixes unable to join error

* Core/Pet: Fixes pet names showing as unknown

* Core/Movement: Fixes scripts that use stuff like SetOrientation after a waypoint,  why are we checking if we are charging? when pointmovement can be used to just move points.

* Fixes server shutdowns

* Core/Talents: Fixes switching talents and specs

* Core/Battleground: Simplified Battleground.SendPacketToTeam

* Core/Packets: Fix null exception on string

* Core/Talents: Really fix talents this time.

Core/Misc: Misc fixes

* Core/Database: Get rid of exception and just print the file the errored.

* Core/Databse: Really fix the error when applying files that do not change any rows.

* Core/Gameobject: fixed sometimes not found linked trap

* Core/Spell: spells with ignore hit result should pierce damage immunities

* Core/Spell: register items received through SPELL_EFFECT_SUMMON_CHANGE_ITEM for quest credit

* Core/Spell: restore old proc system behaviour on auras self proc.

* Core/Entities: better support of absorb, resist and block procs

* Core/Scripts: Fixes ValidateSpellScripts, and type mismatch of parameters

* Core/Spells: Fixes some spells being negative when they should be positive

* Core/Battleground: Dropped battleground texts from trinity_string and replaced them with proper BroadcastText

* Misc fixes and missed sql updates

* Core/PacketIO: Fixed SMSG_SPELL_NON_MELEE_DAMAGE_LOG struct

+++ Source/Game/Combat/ThreatManager.cs
@@ -1,5 +1,5 @@
 ﻿/*
- * Copyright (C) 2012-2017 CypherCore <http://github.com/CypherCore>
+ * Copyright (C) 2012-2018 CypherCore <http://github.com/CypherCore>
  * 
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -381,7 +381,7 @@ namespace Game.Combat
                 Contract.Assert(target);                                     // if the ref has status online the target must be there !
 
                 // some units are prefered in comparison to others
-                if (!noPriorityTargetFound && (target.IsImmunedToDamage(attacker.GetMeleeDamageSchoolMask()) || target.HasNegativeAuraWithInterruptFlag((uint)SpellAuraInterruptFlags.TakeDamage)))
+                if (!noPriorityTargetFound && (target.IsImmunedToDamage(attacker.GetMeleeDamageSchoolMask()) || target.HasNegativeAuraWithInterruptFlag(SpellAuraInterruptFlags.TakeDamage)))
                 {
                     if (i != threatList.Count - 1)
                     {
