+++ Source/Game/DataStorage/Structs/S_Records.cs
@@ -1,5 +1,5 @@
 ﻿/*
- * Copyright (C) 2012-2017 CypherCore <http://github.com/CypherCore>
+ * Copyright (C) 2012-2018 CypherCore <http://github.com/CypherCore>
  * 
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -165,7 +165,7 @@ namespace Game.DataStorage
         public uint ProcTypeMask;
         public uint ProcCategoryRecovery;
         public ushort CumulativeAura;
-        public byte SpellProcsPerMinuteID;
+        public ushort SpellProcsPerMinuteID;
         public byte DifficultyID;
         public byte ProcChance;
     }
